refactor: services into directories
