+++ lib/flutter_offline_sync.dart
@@ -3,4 +3,4 @@ export 'src/generator/entity_registry.dart';
 export 'src/flutter_sync.dart';
 export 'src/ui/configurations_view.dart';
 export 'src/flutter_sync_util.dart' show FlutterSyncProvider;
-export 'src/data/services/data_adapter.dart';
+export 'src/data/services/data/data_adapter.dart';
