+++ lib/src/data/services/data/data_adapter_interface.dart
