+++ lib/src/data/services/syncroniser_service.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_offline_sync/flutter_offline_sync.dart';
-import 'package:flutter_offline_sync/src/data/interfaces/data_syncroniser.dart';
+import 'package:flutter_offline_sync/src/data/services/syncroniser/data_syncroniser_interface.dart';
 import 'package:flutter_offline_sync/src/data/models/data_entity.dart';
 import 'package:flutter_offline_sync/src/utils/logger.dart';
